+++ angel-ps/core/src/main/java/com/tencent/angel/psagent/matrix/oplog/cache/DenseDoubleRowUpdateSplit.java
@@ -51,15 +51,15 @@ public class DenseDoubleRowUpdateSplit extends RowUpdateSplit {
   @Override
   public void serialize(ByteBuf buf) {
     super.serialize(buf);
-    buf.writeInt(end - start);
+    buf.writeInt((int)(end - start));
     LOG.debug("double size = " + (end - start));
-    for (int i = start; i < end; i++) {
+    for (int i = (int)start; i < end; i++) {
       buf.writeDouble(values[i]);
     }
   }
 
   @Override
   public int bufferLen() {
-    return super.bufferLen() + size() * 8;
+    return super.bufferLen() + (int)size() * 8;
   }
 }
